fix: update transaction status change to use correct transaction ID in cancel_transaction handler

+++ src/db/transaction_repository.rs
@@ -105,7 +105,7 @@ impl TransactionRepository {
             SET status = ?
             WHERE id = ?
             "#,
-            status as i16,
+            status,
             transaction_id
         )
         .execute(&self.pool)
